Add --list and --info flags for pools

+++ examples/noso-go.bat
@@ -1,4 +1,5 @@
 :loop
+@echo off
 
 Rem #################################
 Rem ## Begin of user-editable part ##
@@ -8,21 +9,18 @@ REM Setting CPU to 0 will make it automatically use MAXCORES / 2
 REM On my 4 core, 8 threaded i7 that means it will use all 4 physical cores
 REM If you set this to something other than 0, it should be to however many
 REM physical cores you have, not logical
+set "POOL=yzpool"
 set "CPU=2"
 set "WALLET=N2RUEEpVEyF9fgmQg6HKcrwkm4MERDx"
-set "ADDRESS=noso.dukedog.io"
-set "PORT=8082"
-set "PASSWORD=duke"
 
 Rem #################################
 Rem ##  End of user-editable part  ##
 Rem #################################
-@echo off
 setlocal enableDelayedExpansion
 
-taskkill /F /im noso-go.exe
-timeout 3
+tasklist /FI "IMAGENAME eq noso-go.exe" 2>NUL | find /I /N "noso-go.exe">NUL
+if "%ERRORLEVEL%"=="0" taskkill /F /im noso-go.exe
 
-noso-go.exe mine --address !ADDRESS! --port !PORT! --password !PASSWORD! --wallet !WALLET! --cpu !CPU!
+noso-go.exe mine pool !POOL! --wallet !WALLET! --cpu !CPU!
 timeout 10
 goto loop
